requirements.txt
aioice #==0.9.0
aioquic #==1.2.0
aiortc #==1.9.0
annotated-types #==0.7.0
anthropic #==0.46.0
anyio #==4.4.0
attrs #==25.1.0
av #==12.3.0
blinker #==1.9.0
cachetools #==5.5.0
certifi #==2024.8.30
cffi #==1.17.1
charset-normalizer #==3.3.2
click #==8.1.7
colorama #==0.4.6
contourpy #==1.3.0
cryptography #==43.0.1
cycler #==0.12.1
distro #==1.9.0
dnspython #==2.6.1
exceptiongroup #==1.2.2
fastapi #==0.114.1
fastdotcom #==0.0.3
filelock #==3.16.0
Flask #==3.1.0
fonttools #==4.53.1
fsspec #==2024.9.0
ftfy #==6.3.1
gevent #==24.11.1
google-ai-generativelanguage #==0.6.15
google-api-core #==2.19.2
google-api-python-client #==2.161.0
google-auth #==2.34.0
google-auth-httplib2 #==0.2.0
google-cloud-core #==2.4.1
google-cloud-storage #==2.18.2
google-cloud-vision #==3.7.4
google-crc32c #==1.6.0
google-generativeai #==0.8.4
google-resumable-media #==2.7.2
googleapis-common-protos #==1.65.0
greenlet #==3.1.1
grpcio #==1.71.0
grpcio-status #==1.66.1
grpcio-tools #==1.71.0
h11 #==0.14.0
httpcore #==1.0.5
httplib2 #==0.22.0
httpx #==0.27.2
huggingface-hub #==0.28.1
idna #==3.8
ifaddr #==0.2.0
imageio #==2.35.1
imutils #==0.5.4
inexactsearch #==1.0.2
itsdangerous #==2.2.0
Jinja2 #==3.1.4
jiter #==0.8.2
joblib #==1.4.2
kiwisolver #==1.4.7
lazy_loader #==0.4
Levenshtein #==0.26.1
MarkupSafe #==2.1.5
matplotlib #==3.9.2
mpmath #==1.3.0
networkx #==3.3
numpy #==1.26.4
open_clip_torch #==2.26.1
opencv-python #==4.10.0.84
packaging #==24.1
paho-mqtt #==2.1.0
pandas #==2.2.2
pillow #==10.4.0
proto-plus #==1.24.0
protobuf #==6.30.0
psutil #==6.0.0
py-cpuinfo #==9.0.0
pyasn1 #==0.6.1
pyasn1_modules #==0.4.1
pycparser #==2.22
pydantic #==2.9.1
pydantic_core #==2.23.3
pyee #==12.0.0
pylibsrtp #==0.10.0
pylsqpack #==0.3.19
pymongo #==4.8.0
pyOpenSSL #==24.2.1
pyparsing #==3.1.4
pyspellchecker #==0.8.1
python-dateutil #==2.9.0.post0
python-dotenv #==1.0.1
python-Levenshtein #==0.26.1
pytz #==2024.2
PyYAML #==6.0.2
rapidfuzz #==3.9.7
regex #==2024.11.6
requests #==2.32.3
rsa #==4.9
safetensors #==0.5.2
scikit-image #==0.24.0
scikit-learn #==1.5.2
scipy #==1.14.1
seaborn #==0.13.2
sentence-transformers #==3.0.1
service-identity #==24.2.0
setuptools #==75.8.0
silpa_common #==0.3
six #==1.16.0
sniffio #==1.3.1
soundex #==1.1.3
spellchecker #==0.4
starlette #==0.38.5
sympy #==1.13.2
thefuzz #==0.22.1
threadpoolctl #==3.5.0
tifffile #==2024.8.30
timm #==1.0.14
tokenizers #==0.21.0
torch #==2.4.1
torchvision #==0.19.1
tqdm #==4.66.5
transformers #==4.49.0
typing_extensions #==4.12.2
tzdata #==2024.1
ultralytics #==8.2.92
ultralytics-thop #==2.0.6
uritemplate #==4.1.1
urllib3 #==2.2.3
uuid #==1.30
uvicorn #==0.30.6
wcwidth #==0.2.13
websocket #==0.2.1
websockets #==15.0.1
Werkzeug #==3.1.3
wsproto #==1.2.0
zope.event #==5.0
zope.interface #==7.2
